+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - missing fields for ``make_invoice`` response
 - new `fees_paid` field in ``pay_keysend``/``multi_pay_keysend``
 
+### Fixed
+- Don't ignore `offset` parameter in ``list_transactions``
+
 ## [0.1.5] 2025-07-24
 
 ### Changed